… openbas-front/src/actions/exercises/exercise-action.ts
… openbas-front/src/actions/exercises/exercise-teams-action.ts
… openbas-front/src/actions/scenarios/scenario-actions.ts
… openbas-front/src/actions/scenarios/scenario-teams-action.ts
… openbas-front/src/actions/teams/team-actions.ts
… openbas-front/src/admin/components/assets/endpoints/EndpointsDialogAdding.tsx
… openbas-front/src/admin/components/common/Context.ts
… openbas-front/src/admin/components/common/ToolBar.js
… openbas-front/src/admin/components/components/teams/AddTeams.tsx
… openbas-front/src/admin/components/components/teams/CreateTeam.tsx
… openbas-front/src/admin/components/components/teams/UpdateTeams.tsx
… openbas-front/src/admin/components/scenarios/scenario/ScenarioDefinition.tsx
… openbas-front/src/admin/components/scenarios/scenario/teams/ScenarioTeams.tsx
… openbas-front/src/admin/components/simulations/simulation/teams/ExerciseTeams.tsx
… openbas-front/src/admin/components/teams/Teams.tsx
… openbas-front/src/components/ItemTags.js
… openbas-front/src/components/common/SelectList.tsx
… openbas-front/src/components/common/queryable/filter/useRetrieveOptions.tsx
… openbas-front/src/components/common/queryable/filter/useSearchOptions.tsx
… openbas-front/src/utils/Localization.js
… openbas-front/src/utils/api-types.d.ts
… openbas-model/src/main/java/io/openbas/database/model/Team.java
… openbas-model/src/main/java/io/openbas/database/raw/RawPaginationTeam.java
… openbas-model/src/main/java/io/openbas/database/specification/TeamSpecification.java