### Proposed changes

* Change the query mechanism to use Criteria Builder instead of JPA

### Related issues

* https://github.com/OpenBAS-Platform/openbas/issues/1409

### Metrics

Page : Scenario ID - Defeinition
URL : /admin/scenarios/SCENARIO_ID/definition
Api Path : /api/teams
Mechanism | Jaeger time
-- | --
JPA → RawPaginationTeam with pagination (buildPaginationJPA) | 90 < 110 ms
CriteriaBuilder → TeamOutput with pagination (buildPaginationCriteriaBuilder) | 20 < 25 ms

Page : Teams
URL : /admin/teams/teams
Api Path : /api/teams/search
Mechanism | Jaeger time
-- | --
JPA → RawPaginationTeam with pagination (buildPaginationJPA) | 90 < 110 ms
CriteriaBuilder → TeamOutput with pagination (buildPaginationCriteriaBuilder) | 50 < 60 ms