@MarineLeM 2024-09-26
Is it normal that I don’t have the same behavior as in the master branch? When adding a team to an exercise, I can no longer create teams, because these button doesn't appears anymore : 

![image](https://github.com/user-attachments/assets/b3d0c27f-e281-4316-bd4e-d1739e855410)