@MarineLeM 2024-09-26
i found another different behavior from 'master' branch. When you already have teams assigned to an exercise and click on 'add team', on 'master' branch nothing appears in the right square (see the image below to compare with the new behavior implemented). This component is intended solely for adding new teams, not for removing them from the exercise.

If this new behavior is intentional, please be cautious: when you attempt to remove a chip and then click 'add', you receive a message indicating that everything has been updated successfully, but in reality, the team remains assigned.
![image](https://github.com/user-attachments/assets/e8f1e6c0-6a9e-44af-9fbd-fd125cded8e9)