@RomuDeuxfois 2024-09-26
I forget this button... It should be there.

> Is it normal that I don’t have the same behavior as in the master branch? When adding a team to an exercise, I can no longer create teams, because these button doesn't appears anymore :
> 
> ![image](https://private-user-images.githubusercontent.com/17729566/371081571-b3d0c27f-e281-4316-bd4e-d1739e855410.png?jwt=eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJpc3MiOiJnaXRodWIuY29tIiwiYXVkIjoicmF3LmdpdGh1YnVzZXJjb250ZW50LmNvbSIsImtleSI6ImtleTUiLCJleHAiOjE3MjczNDgyNDQsIm5iZiI6MTcyNzM0Nzk0NCwicGF0aCI6Ii8xNzcyOTU2Ni8zNzEwODE1NzEtYjNkMGMyN2YtZTI4MS00MzE2LWJkNGUtZDE3MzllODU1NDEwLnBuZz9YLUFtei1BbGdvcml0aG09QVdTNC1ITUFDLVNIQTI1NiZYLUFtei1DcmVkZW50aWFsPUFLSUFWQ09EWUxTQTUzUFFLNFpBJTJGMjAyNDA5MjYlMkZ1cy1lYXN0LTElMkZzMyUyRmF3czRfcmVxdWVzdCZYLUFtei1EYXRlPTIwMjQwOTI2VDEwNTIyNFomWC1BbXotRXhwaXJlcz0zMDAmWC1BbXotU2lnbmF0dXJlPTY5YzNiMjE5OWI2ODI1MjYwOGQ3ZjkxZjE1MjJhZGQ4MTEzZDhjYTdlOGRkNmNjMTFkZWQyMDlkNmRiZTFkOGUmWC1BbXotU2lnbmVkSGVhZGVycz1ob3N0In0.JgHwjlsaOPqvRcOZwaMUGqtK-CDe-zDbhRz-FblcK2U)

You are right.
I will change the behavior to have a replace method insteaf of an add method.
To be able to remove multiple teams easily.
Please, test it and tell me if it's better.

> i found another different behavior from 'master' branch. When you already have teams assigned to an exercise and click on 'add team', on 'master' branch nothing appears in the right square (see the image below to compare with the new behavior implemented). This component is intended solely for adding new teams, not for removing them from the exercise.
> 
> If this new behavior is intentional, please be cautious: when you attempt to remove a chip and then click 'add', you receive a message indicating that everything has been updated successfully, but in reality, the team remains assigned. ![image](https://private-user-images.githubusercontent.com/17729566/371086135-e8f1e6c0-6a9e-44af-9fbd-fd125cded8e9.png?jwt=eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJpc3MiOiJnaXRodWIuY29tIiwiYXVkIjoicmF3LmdpdGh1YnVzZXJjb250ZW50LmNvbSIsImtleSI6ImtleTUiLCJleHAiOjE3MjczNDgyNDQsIm5iZiI6MTcyNzM0Nzk0NCwicGF0aCI6Ii8xNzcyOTU2Ni8zNzEwODYxMzUtZThmMWU2YzAtNmE5ZS00NGFmLTlmYmQtZmQxMjVjZGVkOGU5LnBuZz9YLUFtei1BbGdvcml0aG09QVdTNC1ITUFDLVNIQTI1NiZYLUFtei1DcmVkZW50aWFsPUFLSUFWQ09EWUxTQTUzUFFLNFpBJTJGMjAyNDA5MjYlMkZ1cy1lYXN0LTElMkZzMyUyRmF3czRfcmVxdWVzdCZYLUFtei1EYXRlPTIwMjQwOTI2VDEwNTIyNFomWC1BbXotRXhwaXJlcz0zMDAmWC1BbXotU2lnbmF0dXJlPTUzMmI2ZjQ0NWIyMjhmOWY5ZTg0ZmZhNDE2YzNhN2M0YmZiZjQ0NTU2MzNhYTE3OWU0NWJiOTkwMGVmNDUxODEmWC1BbXotU2lnbmVkSGVhZGVycz1ob3N0In0._INMfHR9oCQ-nYerAihHMdG9R3-8iKQqcIWvAv_tEqE)