@MarineLeM 2024-09-26
The 'remove from the context' button does not work anympre
![image](https://github.com/user-attachments/assets/27262a7d-ee6c-4ef1-8111-cfe7be30a1b8)