@RomuDeuxfois 2024-09-26
Fix

> The 'remove from the context' button does not work anympre ![image](https://private-user-images.githubusercontent.com/17729566/371148042-27262a7d-ee6c-4ef1-8111-cfe7be30a1b8.png?jwt=eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJpc3MiOiJnaXRodWIuY29tIiwiYXVkIjoicmF3LmdpdGh1YnVzZXJjb250ZW50LmNvbSIsImtleSI6ImtleTUiLCJleHAiOjE3MjczNTczMzcsIm5iZiI6MTcyNzM1NzAzNywicGF0aCI6Ii8xNzcyOTU2Ni8zNzExNDgwNDItMjcyNjJhN2QtZWU2Yy00ZWYxLTgxMTEtY2ZlN2JlMzBhMWI4LnBuZz9YLUFtei1BbGdvcml0aG09QVdTNC1ITUFDLVNIQTI1NiZYLUFtei1DcmVkZW50aWFsPUFLSUFWQ09EWUxTQTUzUFFLNFpBJTJGMjAyNDA5MjYlMkZ1cy1lYXN0LTElMkZzMyUyRmF3czRfcmVxdWVzdCZYLUFtei1EYXRlPTIwMjQwOTI2VDEzMjM1N1omWC1BbXotRXhwaXJlcz0zMDAmWC1BbXotU2lnbmF0dXJlPTRiOTJhYjNjYThiYTc5YzQ0OTdkNDU5YTU2Mzc1ODNjNDk1MWFkNjljNzY5YzZlZmE4YTg5Y2Q4M2YxNTBmYmUmWC1BbXotU2lnbmVkSGVhZGVycz1ob3N0In0.jOxG-_9m97XqNPnVoipbhcuQFez9Sx92M_jcGfXiGnI)